change VAOS mount point from /services/vaos to /v0/vaos
