--- modules/vaos/app/controllers/vaos/v0/systems_controller.rb
@@ -1,18 +1,0 @@
-# frozen_string_literal: true
-
-module VAOS
-  module V0
-    class SystemsController < ApplicationController
-      def index
-        response = systems_service.get_systems(current_user)
-        render json: VAOS::SystemSerializer.new(response)
-      end
-
-      private
-
-      def systems_service
-        VAOS::SystemsService.new
-      end
-    end
-  end
-end
